# backend/api-first/openapi.yaml
openapi: 3.1.0
info:
  title: KOSMO API
  version: "0.1.0"
  description: >-
    API de KOSMO - plataforma low-code de spec-driven development asistida por IA.
    Ciclo completo: Descubrimiento -> Caracteristicas -> Requisitos -> Modelo -> Codigo.

    Convenciones:
    - Autenticacion PKCE (RFC 7636) con JWT RS256 (bearerAuth).
    - Errores RFC 7807 (application/problem+json) con type URN, trace_id y violations.
    - Fechas ISO-8601 UTC con sufijo Z.
    - Las fases siguen el orden Descubrimiento -> Caracteristicas -> Requisitos -> Modelo.
      La consistencia solo se evalua hacia la derecha (downstream).

    Los endpoints de generacion de codigo estan marcados con x-status: planned;
    el frontend puede integrarse contra ellos mientras el backend los implementa.
  contact:
    name: KOSMO Team
    email: dev@kosmo.app
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/consistency/evaluate:
    post:
      tags: [consistency]
      summary: Evaluar consistencia entre fases (solo downstream)
      description: >-
        Evalua el impacto de cambios sobre las fases a la derecha segun el orden
        Descubrimiento -> Caracteristicas -> Requisitos -> Modelo. Si se solicita
        evaluar hacia la izquierda (upstream), devuelve 422.
      operationId: consistency_evaluate
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateConsistencyRequestView"
      responses:
        "200":
          description: Reporte de consistencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsistencyReportView"
        "422":
          description: Direccion de evaluacion no permitida (upstream)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{project_id}/consistency/evaluate/stream:
    post:
      tags: [consistency]
      summary: Evaluar consistencia con streaming SSE
      description: >-
        Igual que /evaluate pero con progreso en tiempo real (SSE). Aplica la
        misma restriccion downstream-only.
      operationId: consistency_evaluate_stream
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateConsistencyRequestView"
      responses:
        "200":
          description: Stream SSE con progreso y reporte final
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ConsistencyStreamEvent"
        "422":
          description: Direccion de evaluacion no permitida (upstream)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{project_id}/consistency/apply:
    post:
      tags: [consistency]
      summary: Aplicar impactos de consistencia
      description: >-
        Aplica las acciones sugeridas sobre requisitos, caracteristicas y
        diagramas (artefactos downstream).
      operationId: consistency_apply
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyImpactsRequestView"
      responses:
        "200":
          description: Resultado de aplicacion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyImpactsResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ══════════════════════════════ SCHEMAS ══════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 emitido por /auth/token o /auth/refresh
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador del proyecto (prj_ + ULID)
...
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    ProblemDetail:
      type: object
      description: Error RFC 7807 (application/problem+json)
      properties:
        type:
          type: string
          description: URN del error (ej. urn:kosmo:projects:not-found)
          example: urn:kosmo:projects:not-found
        title:
          type: string
          description: Mensaje corto
        status:
          type: integer
          description: Codigo HTTP
        detail:
          type: string
          description: Descripcion larga
        instance:
          type: string
          description: Ruta que fallo
        trace_id:
          type: string
          description: Identificador de correlacion (X-Trace-Id)
        violations:
          type: array
          items:
            type: object
            properties:
              loc:
                type: string
                description: Ruta al campo invalido
              msg:
                type: string
              input:
                description: Valor recibido
      required: [type, title, status, detail]

    # ── Auth ──
...
    ChangeInputView:
      type: object
      properties:
        section:
          type: string
        diff_before:
          type: string
        diff_after:
          type: string
        description:
          type: string
          default: ""
      required: [section, diff_before, diff_after]
    EvaluateConsistencyRequestView:
      type: object
      additionalProperties: false
      properties:
        phase_origin:
          type: string
          description: Fase de origen (descubrimiento, caracteristicas, requisitos, modelo)
        phase_destination:
          type: string
          nullable: true
          description: >-
            Fase destino a evaluar. Debe estar a la derecha de phase_origin
            (Descubrimiento -> Caracteristicas -> Requisitos -> Modelo).
            Si se omite, se evaluan todas las fases a la derecha.
        changes:
          type: array
          items:
            $ref: "#/components/schemas/ChangeInputView"
      required: [phase_origin, changes]
    ImpactItemView:
      type: object
      properties:
        phase:
          type: string
        artifact_id:
          type: string
        artifact_type:
          type: string
        artifact_label:
          type: string
        section:
          type: string
          default: ""
        rationale:
          type: string
          default: ""
        diff_suggestion:
          type: object
          nullable: true
      required: [phase, artifact_id, artifact_type, artifact_label]
    ConsistencyReportView:
      type: object
      properties:
        id:
          type: string
          description: ID del reporte (cnr_ + ULID)
        phase_origin:
          type: string
        own_changes:
          type: array
          items:
            $ref: "#/components/schemas/ChangeInputView"
        downstream_impact:
          type: array
          description: Impacto en fases a la derecha
          items:
            $ref: "#/components/schemas/ImpactItemView"
        created_at:
          type: string
          format: date-time
      required: [id, phase_origin, own_changes, downstream_impact, created_at]
    ConsistencyStreamEvent:
      oneOf:
        - type: object
          description: Progreso de evaluacion
          properties:
            type:
              type: string
              const: progress
            message:
              type: string
          required: [type, message]
        - type: object
          description: Reporte final
          properties:
            type:
              type: string
              const: report
            report:
              $ref: "#/components/schemas/ConsistencyReportView"
          required: [type, report]
    ApplyImpactsRequestView:
      type: object
      additionalProperties: false
      properties:
        impacts:
          type: array
          items:
            type: object
            properties:
              artifact_type:
                type: string
                description: EARSRequirement, Feature, ActivityDiagram, DiscoveryDocument
              target_id:
                type: string
              action:
                type: string
                description: update o delete
              field:
                type: string
                default: description
              before:
                type: string
                default: ""
              after:
                type: string
                default: ""
            required: [artifact_type, target_id]
      required: [impacts]
    ApplyImpactsResponse:
      type: object
      properties:
        applied:
          type: array
          items:
            type: object
            properties:
              target_id:
                type: string
              artifact_type:
                type: string
            required: [target_id, artifact_type]
        failed:
          type: array
          items:
            type: object
            properties:
              target_id:
                type: string
              artifact_type:
                type: string
              reason:
                type: string
            required: [target_id, artifact_type, reason]
      required: [applied, failed]

    # ── Implementation (codegen, planificada) ──
...